requirements.txt
streamlit
pandas
openpyxl
requests
pillow
pillow-heif
PyGithub